+++ sdk/cosmos/azure_data_cosmos/examples/cosmos/create.rs
@@ -97,7 +97,7 @@ impl CreateCommand {
                 println!("Created item successfully");
 
                 if show_updated {
-                    let created: serde_json::Value = response.into_raw_body().json().await?;
+                    let created: serde_json::Value = response.into_raw_body().json()?;
                     println!("Newly created item:");
                     println!("{:#?}", created);
                 }
@@ -115,11 +115,7 @@ impl CreateCommand {
                     ..Default::default()
                 });
 
-                let db = client
-                    .create_database(&id, options)
-                    .await?
-                    .into_body()
-                    .await?;
+                let db = client.create_database(&id, options).await?.into_body()?;
                 println!("Created database:");
                 println!("{:#?}", db);
                 Ok(())
@@ -163,8 +159,7 @@ impl CreateCommand {
                     .database_client(&database)
                     .create_container(properties, options)
                     .await?
-                    .into_body()
-                    .await?;
+                    .into_body()?;
                 println!("Created container:");
                 println!("{:#?}", container);
                 Ok(())