+++ sdk/eventhubs/azure_messaging_eventhubs_checkpointstore_blob/src/checkpoint_store.rs
@@ -238,7 +238,7 @@ impl CheckpointStore for BlobCheckpointStore {
         };
 
         while let Some(blob) = blobs.try_next().await? {
-            let blob_body = blob.into_body().await?;
+            let blob_body = blob.into_body()?;
             debug!("Blob body: {blob_body:?}, {:?}", blob_body.container_name);
             for blob in blob_body.segment.blob_items.iter() {
                 let mut checkpoint = checkpoint.clone();
@@ -307,7 +307,7 @@ impl CheckpointStore for BlobCheckpointStore {
         };
 
         while let Some(blob) = blobs.try_next().await? {
-            let blob_body = blob.into_body().await?;
+            let blob_body = blob.into_body()?;
             debug!("Blob body: {blob_body:?}, {:?}", blob_body.container_name);
             for blob in blob_body.segment.blob_items.iter() {
                 let mut ownership = ownership.clone();