+++ sdk/keyvault/azure_security_keyvault_secrets/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Breaking Changes
 
+- Client methods that return a `Response<T>>` asynchronously buffer the entire model within the internal pipeline, so `into_body()` and other methods on the response are no longer async.
+
 ### Bugs Fixed
 
 ### Other Changes