+++ sdk/storage/azure_storage_queue/CHANGELOG.md
@@ -5,3 +5,7 @@
 ### Features Added
 
 * Initial supported release.
+
+### Breaking Changes
+
+- Client methods that return a `Response<T>>` asynchronously buffer the entire model within the internal pipeline, so `into_body()` and other methods on the response are no longer async.